✨ feat: 重構儀表板功能，整合 WakaTime 和 YouTube 數據，移除不必要的檔案
