+++ src/lib/dashboard/types.ts
@@ -1,0 +1,10 @@
+type DashboardFeature = {
+  title: string;
+  description: string;
+  link: string;
+  image: React.ReactNode;
+  source: string;
+  color: Record<string, string>;
+};
+
+export type { DashboardFeature };
